enhance GalleryImageModel.kt
This commit prepares the app to welcome a Fragment aimed at showing an image's details

+++ app/src/main/java/com/example/galleriquii/dto/RedditResponseDto.kt
@@ -430,7 +430,7 @@ internal class ChildrenData {
 
     @SerializedName("created_utc")
     @Expose
-    var createdUtc: Int? = null
+    var createdUtc: Long? = null
 
     @SerializedName("num_crossposts")
     @Expose
